refactor(repository, meetup): meetup repository 통합
- @Nullable을 사용해 매개변수에서 Optional 사용 제거
